layout/forms/test/test_bug378670.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=378670
-->
<head>
  <title>Test for Bug 378670</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=378670">Mozilla Bug 378670</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
Clicking on the select should not crash Mozilla
<select id="select">
<option>1</option>
<option>2</option>
</select>

<pre id="test">
<script>
document.body.addEventListener('popupshowing', function(e) {e.target.remove() }, true);
</script>
<script type="application/javascript">

/** Test for Bug 378670 **/

SimpleTest.waitForExplicitFinish();
SimpleTest.requestFlakyTimeout("untriaged");

function clickit() {
  var select = document.getElementById('select');

  sendMouseEvent({type:'mousedown'}, select);
  sendMouseEvent({type:'mouseup'}, select);
  sendMouseEvent({type:'click'}, select);

  setTimeout(finish, 200);
}

window.addEventListener('load', clickit);

function finish()
{
  ok(true, "This is a mochikit version of a crash test. To complete is to pass.");
  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
